requirements.txt
google-generativeai==0.8.3
supabase==2.7.4
aiohttp==3.9.5
asyncio==3.4.3
requests==2.31.0
python-dotenv==1.0.1
pandas==2.2.2
numpy==1.26.4
python-dateutil==2.9.0
